refactor(courses): address deletion review findings
