+++ apps/frontend-manage/src/lib/utils/graphqlErrors.ts
@@ -1,0 +1,20 @@
+export function getGraphQLErrorCode(error: unknown): string | undefined {
+  if (!error || typeof error !== 'object') return undefined
+
+  const extensions = (error as { extensions?: { code?: unknown } }).extensions
+  if (typeof extensions?.code === 'string') return extensions.code
+
+  const graphQLErrors = (error as { graphQLErrors?: unknown[] }).graphQLErrors
+  for (const graphQLError of graphQLErrors ?? []) {
+    const code = getGraphQLErrorCode(graphQLError)
+    if (code) return code
+  }
+
+  const errors = (error as { errors?: unknown[] }).errors
+  for (const nestedError of errors ?? []) {
+    const code = getGraphQLErrorCode(nestedError)
+    if (code) return code
+  }
+
+  return undefined
+}
